+++ test/test_helper.js
@@ -15,7 +15,7 @@
 // ****************************************************************************
 
 const Aerospike = require('../lib/aerospike')
-const info = require('../lib/info')
+const Info = require('../lib/info')
 const options = require('./util/options')
 const expect = require('expect.js')
 const deasync = require('deasync')
@@ -65,30 +65,44 @@ UDFHelper.prototype.remove = function (filename, done) {
 
 function IndexHelper (client) {
   this.client = client
-  this.createStringIndex = deasync(client.createStringIndex).bind(client)
-  this.createIntegerIndex = deasync(client.createIntegerIndex).bind(client)
-  this.indexCreateWait = deasync(client.indexCreateWait).bind(client)
+  this.createIndex = deasync(client.createIndex).bind(client)
   this.indexRemove = deasync(client.indexRemove).bind(client)
+  this.info = deasync(client.info).bind(client)
 }
 
-IndexHelper.prototype.create = function (index_name, set_name, bin_name, index_type, done) {
+IndexHelper.prototype.create = function (indexName, setName, binName, dataType, indexType) {
   var index = {
     ns: options.namespace,
-    set: set_name,
-    bin: bin_name,
-    index: index_name
+    set: setName,
+    bin: binName,
+    index: indexName,
+    type: indexType || Aerospike.indexType.DEFAULT,
+    datatype: dataType
   }
-  switch (index_type.toLowerCase()) {
-    case 'string': this.createStringIndex(index); break
-    case 'integer': this.createIntegerIndex(index); break
-  }
-  this.indexCreateWait(index.ns, index.index, 100)
-  if (done) done()
+  var task = this.createIndex(index)
+  deasync(task.waitUntilDone).bind(task)(100)
 }
 
-IndexHelper.prototype.remove = function (index_name, done) {
-  this.indexRemove(options.namespace, index_name, {})
-  if (done) done()
+IndexHelper.prototype.remove = function (indexName) {
+  this.indexRemove(options.namespace, indexName, {})
+}
+
+IndexHelper.prototype.exists = function (indexName) {
+  var sindex = 'sindex/' + options.namespace + '/' + indexName
+  var exists = false
+  var attempts = 0
+  while (attempts < 5) {
+    attempts++
+    this.info(sindex, function (err, info) {
+      if (err) throw err
+      var indexStats = Info.parseInfo(info)[sindex]
+      var noIndexErr = (typeof indexStats === 'string') && indexStats.indexOf('FAIL:201:NO INDEX') >= 0
+      exists = exists || !noIndexErr
+    })
+    if (exists) return exists
+    deasync.sleep(5)
+  }
+  return false
 }
 
 function ServerInfoHelper () {
@@ -110,7 +124,7 @@ ServerInfoHelper.prototype.fetch_info = function () {
   var done = false
   client.info('features', function (err, result) {
     if (err) throw err
-    self.features = info.parseInfo(result)['features']
+    self.features = Info.parseInfo(result)['features']
   }, function () { done = true })
   deasync.loopWhile(function () { return !done })
 }
@@ -121,7 +135,7 @@ ServerInfoHelper.prototype.fetch_namespace_config = function (ns) {
   var nsKey = 'namespace/' + ns
   client.info(nsKey, function (err, result) {
     if (err) throw err
-    self.nsconfig = info.parseInfo(result)[nsKey]
+    self.nsconfig = Info.parseInfo(result)[nsKey]
   }, function () { done = true })
   deasync.loopWhile(function () { return !done })
 }